new website from skratch
This commit radically changes the website, as not it will be based on a
free template that can be found at https://html5up.net/ .

The content that is made available with this commit is the redirect page
that points at the italian version of the website, which at the moment
is a copypaste of the english version.
The latter is still to be completed but visually resambles to what it
should become.